backend/sale_platform/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.6
python-decouple==3.8
drf-yasg==1.21.7  # برای Swagger documentation